requirements.txt
streamlit==1.28.2
langchain==0.0.338
openai==1.3.3
tiktoken==0.5.1

Pillow==10.1.0